requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.0
attrs==24.2.0
backoff==2.2.1
basicsr==1.4.2
blinker==1.8.2
branca==0.8.0
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
chess==1.11.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Deprecated==1.2.14
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
fire==0.7.0
Flask==3.0.3
flatbuffers==24.3.25
folium==0.18.0
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
grpcio==1.68.1
gTTS==2.5.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
narwhals==1.4.1
networkx==3.3
numba==0.60.0
numpy==2.0.1
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
openai-whisper==20240930
opencage==3.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
phonenumbers==8.13.49
pikepdf==9.4.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==5.27.3
psutil==6.1.0
pyarrow==17.0.0
pycparser==2.22
pycryptodomex==3.21.0
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.0
pygame-ce==2.5.2
Pygments==2.18.0
PyMatting==1.1.13
PyMuPDF==1.24.13
pyparsing==3.2.0
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytube3==9.6.4
pytz==2024.1
PyYAML==6.0.2
realesrgan==0.3.0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
rembg==2.0.60
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.11.0
spotipy==2.24.0
streamlit==1.37.1
streamlit-lottie==0.0.5
sympy==1.13.1
tb-nightly==2.19.0a20241206
tenacity==8.5.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tifffile==2024.9.20
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.9.0
yapf==0.43.0
yarl==1.17.1
youtube-dl==2021.12.17
yt-dlp==2024.10.22
